refonte des controller figure/media, ajout de la suppression bannière et modification d'un media, changement pour les vidéos youtube
