+++ src/Controller/FigureController.php
@@ -39,10 +39,7 @@ class FigureController extends AbstractController
         $dataURL = $request->getPathInfo();
 
         $name = substr($dataURL, strpos($dataURL, "show") + 5);
-        //$slugger = new AsciiSlugger();
         $slug = $slugger->slug($name);
-        // $name = urldecode($name);
-        //dd("dans l'URL : $dataURL", " resultat après sluger: $name");
         $figure = $figureRepository->findOneBy(['slug' => $slug]);
 
         if (!$figure) {
@@ -135,128 +132,10 @@ class FigureController extends AbstractController
             return $this->redirectToRoute('app_figure', ['slug' => $figure->getSlug()]);
         }
 
-        //création du formulaire d'ajout d'une video
-        $video = new Media();
-        $videoForm = $this->createForm(VideoFormType::class, $video);
-        $videoForm->handleRequest($request);
-
-        //  Envoie du formulaire d'ajout d'une vidéo
-        if ($videoForm->isSubmitted() && $videoForm->isValid()) {
-            $newVideo = $videoForm->getData();
-
-            // récuparation de l'url
-            $fullstring = $newVideo->getMediaPath();
-            $parsed = $this->getStringBetween($fullstring, 'src="', '"');
-
-            $newVideo->setMediaPath($parsed);
-
-            // Le groupe ici est photo
-            $newVideo->setGroupe(MediaGroupe::video);
-
-            // firstMedia = false car ce n'est pas la bannière
-            $newVideo->setFirstMedia(0);
-
-            // attribution du figure id
-            $newVideo->setFigure($figure);
-
-            $entityManager->persist($newVideo);
-            $entityManager->flush();
-        }
-
-        //création du formulaire d'ajout d'une photo
-        $photo = new Media();
-        $photoForm = $this->createForm(MediaFormType::class, $photo);
-        $photoForm->handleRequest($request);
-        $photoPath = $photoForm->get('media_path')->getData();
-
-        //  Envoie du formulaire d'ajout de photo
-        if ($photoForm->isSubmitted() && $photoForm->isValid()) {
-            $newPhoto = $photoForm->getData();
-            if ($photoPath) {
-                $newFileName = $figure->getId() . '-' . uniqid() . '.' . $photoPath->guessExtension();
-                try {
-                    $photoPath->move(
-                        $this->getParameter('kernel.project_dir') . '/public/uploads/tricksPhoto/',
-                        $newFileName
-                    );
-                } catch (FileException $e) {
-                    return new Response($e->getMessage());
-                }
-                $newPhoto->setMediaPath('uploads/tricksPhoto/' . $newFileName);
-                // Le groupe ici est photo
-                $newPhoto->setGroupe(MediaGroupe::photo);
-                // firstMedia = false car ce n'est pas la bannière
-                $newPhoto->setFirstMedia(0);
-                // attribution du figure id
-                $newPhoto->setFigure($figure);
-
-                $entityManager->persist($newPhoto);
-                $entityManager->flush();
-            }
-        }
-
-        // création du formulaire de modification de la bannière
-        $banner = new Media();
-        $bannerForm = $this->createForm(MediaFormType::class, $banner);
-        $bannerForm->handleRequest($request);
-        $bannerPath = $bannerForm->get('media_path')->getData();
-
-        //  Envoie du formulaire de bannière
-        if ($bannerForm->isSubmitted() && $bannerForm->isValid()) {
-            $newBanner = $bannerForm->getData();
-            if ($bannerPath) {
-                //suppression de l'ancienne bannière
-                if ($figure->getBanner()) {
-                    unlink($figure->getBanner()->getMediaPath());
-                    $entityManager->remove($figure->getBanner());
-                }
-                $newFileName = uniqid() . '.' . $bannerPath->guessExtension();
-                try {
-                    $bannerPath->move(
-                        $this->getParameter('kernel.project_dir') . '/public/uploads/tricksBanner/',
-                        $newFileName
-                    );
-                } catch (FileException $e) {
-                    return new Response($e->getMessage());
-                }
-                $newBanner->setMediaPath('uploads/tricksBanner/' . $newFileName);
-                //  la bannière est forcément une photo
-                $newBanner->setGroupe(MediaGroupe::photo);
-                // firstMedia = true car c'est la bannière
-                $newBanner->setFirstMedia(1);
-                // attribution du figure id
-                $newBanner->setFigure($figure);
-
-                $entityManager->persist($newBanner);
-                $entityManager->flush();
-            }
-        }
-
         return $this->render('figure/edit_figure.html.twig', [
             'figure' => $figure,
             'medias' => $figure->getMedia(),
-            'figureForm' => $figureForm->createView(),
-            'bannerForm' => $bannerForm->createView(),
-            'photoForm' => $photoForm->createView(),
-            'videoForm' => $videoForm->createView()
+            'figureForm' => $figureForm->createView()
         ]);
-    }
-
-    /**
-     * Extrait une chaîne par rapport à une autre chaîne de début et de fin
-     *
-     * @param string $string chaîne de caractère dans laquelel on recherche 
-     * @param string $start chaîne de caractère cherché dans $string pour trouver la position de départ
-     * @param string $end chaîne de caractère cherché dans $string pour trouver la position de fin
-     * @return string chaîne qui se trouve entre $start et $end ou chaîne vide si rien ne match
-     */
-    function getStringBetween($string, $start, $end)
-    {
-        $string = ' ' . $string;
-        $ini = strpos($string, $start);
-        if ($ini == 0) return '';
-        $ini += strlen($start);
-        $len = strpos($string, $end, $ini) - $ini;
-        return substr($string, $ini, $len);
     }
 }
