+++ templates/media/ajout_media_banner.html.twig
@@ -1,0 +1,35 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Ajouter une bannière{% endblock %}
+{% block stylesheets %}
+<link href="../../../assets/style.css" rel="stylesheet">
+{% endblock %}
+
+{% block body %}
+<main class="add-photo-media-page">
+    {# <div class="add-media-form">
+        {{ form_errors(photoForm) }}
+
+        {{ form_start(photoForm, {'attr': {'class': 'addPhotoForm'}}) }}
+            {{ form_label(photoForm.media_path, "Choisissez l'image que vous voulez ajouter :", {
+                'label_attr': {'class': 'addPhotoForm-label'}
+            }) }}
+            {{ form_row(photoForm.media_path, {'attr': {'class': 'addPhotoForm-row'}}) }}
+            <button type="submit" class="register-form-btn">Ajouter</button>
+        {{ form_end(photoForm) }}
+    </div> #}
+
+    <div class="add-media-form">
+        {{ form_errors(bannerForm) }}
+
+        {{ form_start(bannerForm, {'attr': {'class': 'addPhotoForm'}})}}
+            {{ form_label(bannerForm.media_path, "Choisissez l'image qui servira de bannière :", {
+                'label_attr': {'class': 'addPhotoForm-label'}
+            }) }}
+            {{ form_row(bannerForm.media_path, {'attr': {'class': 'addPhotoForm-row'}}) }}
+            <button type="submit" class="register-form-btn">Enregistrer</button>
+        {{ form_end(bannerForm) }}
+    </div>
+
+</main>
+{% endblock %}
