+++ templates/media/ajout_media_video.html.twig
@@ -1,0 +1,23 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Ajouter une vidéo{% endblock %}
+{% block stylesheets %}
+<link href="../../../assets/style.css" rel="stylesheet">
+{% endblock %}
+
+{% block body %}
+<main class="add-photo-media-page">
+    <div class="add-media-form">
+        {{ form_errors(videoForm) }}
+
+        {{ form_start(videoForm, {'attr': {'class': 'addVideo-form'}}) }}
+            {{ form_label(videoForm.media_path, "Le lien de la vidéo (Youtube uniquement) :", {
+                'label_attr': {'class': 'addVideo-form-inputgroup margin'}
+            }) }}
+            {{ form_row(videoForm.media_path, {'attr': {'class': 'addVideo-form-inputgroup'}}) }}
+            <button type="submit" class="register-form-btn">Ajouter</button>
+        {{ form_end(videoForm) }}
+    </div>
+
+</main>
+{% endblock %}
