+++ templates/other/connexionRequired.html.twig
@@ -3,7 +3,7 @@
 {% block title %}Erreur{% endblock %}
 
 {% block body %}
-<div class="">
+<div class="connexionRequired-page">
     <h1>Attention !</h1>
     <p>vous devez être connecté pour aller sur cette page.</p>
     <button class="register-form-btn">
